피드백 반영 - 패키지 구조 변경
entity->domain, repository->persistence
