+++ src/main/java/org/prgrms/deukyun/voucherapp/app/command/CreateVoucherCommand.java
@@ -1,9 +1,9 @@
 package org.prgrms.deukyun.voucherapp.app.command;
 
 import org.prgrms.deukyun.voucherapp.app.console.ConsoleService;
-import org.prgrms.deukyun.voucherapp.domain.voucher.entity.FixedAmountDiscountVoucher;
-import org.prgrms.deukyun.voucherapp.domain.voucher.entity.PercentDiscountVoucher;
-import org.prgrms.deukyun.voucherapp.domain.voucher.entity.Voucher;
+import org.prgrms.deukyun.voucherapp.domain.voucher.domain.FixedAmountDiscountVoucher;
+import org.prgrms.deukyun.voucherapp.domain.voucher.domain.PercentDiscountVoucher;
+import org.prgrms.deukyun.voucherapp.domain.voucher.domain.Voucher;
 import org.prgrms.deukyun.voucherapp.domain.voucher.service.VoucherService;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
