+++ src/main/java/org/prgrms/deukyun/voucherapp/domain/voucher/domain/PercentDiscountVoucher.java
@@ -1,4 +1,4 @@
-package org.prgrms.deukyun.voucherapp.domain.voucher.entity;
+package org.prgrms.deukyun.voucherapp.domain.voucher.domain;
 
 import java.text.MessageFormat;
 import java.util.UUID;
